fix podfile
