+++ Application-Challenge/项目 5-Rethinking-打卡狗/PunchCardTogether/Podfile
@@ -32,3 +32,11 @@ target 'PunchCardTogether' do
   end
 
 end
+
+post_install do |installer|
+  installer.pods_project.targets.each do |target|
+    target.build_configurations.each do |config|
+      config.build_settings['IPHONEOS_DEPLOYMENT_TARGET'] = '14.0'
+    end
+  end
+end
